print the masked token in initialization
