+++ v1/ao/internal/agent/config_builder_test.go
@@ -19,3 +19,15 @@ func TestInitConf(t *testing.T) {
 	Init()
 	assert.True(t, strings.HasSuffix(buffer.String(), "non-default configuration used APPOPTICS_DEBUG_LEVEL=debug\n"))
 }
+
+func TestMaskServiceKey(t *testing.T) {
+	keyPairs := map[string]string{
+		"1234567890abcdef:Go": "1234********cdef:Go",
+		"abc:Go":              "abc:Go",
+		"abcd1234:Go":         "abcd1234:Go",
+	}
+
+	for key, masked := range keyPairs {
+		assert.Equal(t, masked, maskServiceKey(key))
+	}
+}
